move non cli code from main to ferretdb
